requirements.txt
aiofiles
beautifulsoup4
httpx
pyyaml
python-telegram-bot==20.7
openai
urllib3
